+++ src/Catty/ChangeXByNBrick.m
@@ -43,7 +43,7 @@
     return ^{
         NSDebug(@"Performing: %@", self.description);
         double xMov = [self.xMovement interpretDoubleForSprite:self.object];
-        self.object.position = CGPointMake(self.object.position.x+xMov, self.object.position.y);
+        self.object.position = CGPointMake((CGFloat)(self.object.position.x+xMov), self.object.position.y);
 
     };
 }
