+++ src/Catty/NextLookBrick.m
@@ -67,10 +67,10 @@ return ^{
     self.object.texture = texture;
     self.object.currentLook = look;
     if(xScale != 1.0) {
-        self.object.xScale = xScale;
+        self.object.xScale = (CGFloat)xScale;
     }
     if(yScale != 1.0) {
-        self.object.yScale = yScale;
+        self.object.yScale = (CGFloat)yScale;
     }
     };
 }
