+++ src/Catty/SetBrightnessBrick.m
@@ -46,7 +46,7 @@
 {
     return ^{
         NSDebug(@"Performing: %@", self.description);
-        CGFloat brightness = [self.brightness interpretDoubleForSprite:self.object]/100;
+        double brightness = [self.brightness interpretDoubleForSprite:self.object]/100;
         if (brightness > 2) {
             brightness = 1.0f;
         }
@@ -77,7 +77,7 @@
         UIImage *newImage = [UIImage imageWithCGImage:cgimg];
         self.object.currentUIImageLook = newImage;
         self.object.texture = [SKTexture textureWithImage:newImage];
-        self.object.currentLookBrightness = brightness;
+        self.object.currentLookBrightness = (CGFloat)brightness;
         double xScale = self.object.xScale;
         double yScale = self.object.yScale;
         self.object.xScale = 1.0;
@@ -85,10 +85,10 @@
         self.object.size = self.object.texture.size;
         self.object.texture = self.object.texture;
         if(xScale != 1.0) {
-            self.object.xScale = xScale;
+            self.object.xScale = (CGFloat)xScale;
         }
         if(yScale != 1.0) {
-            self.object.yScale = yScale;
+            self.object.yScale = (CGFloat)yScale;
         }
         
         // 4
