+++ src/Catty/SetSizeToBrick.m
@@ -42,11 +42,11 @@
     return ^{
         NSDebug(@"Performing: %@", self.description);
         double sizeInPercent = [self.size interpretDoubleForSprite:self.object];
-        [self.object setXScale:(sizeInPercent/100.0)];
-        [self.object setYScale:(sizeInPercent/100.0)];
+        [self.object setXScale:(CGFloat)(sizeInPercent/100.0f)];
+        [self.object setYScale:(CGFloat)(sizeInPercent/100.0f)];
         //for touch issue
         CGImageRef image = [self.object.currentUIImageLook CGImage];
-        self.object.currentUIImageLook = [UIImage imageWithCGImage:image scale:1/(sizeInPercent/100.0) orientation:UIImageOrientationUp];
+        self.object.currentUIImageLook = [UIImage imageWithCGImage:image scale:(CGFloat)(1.0f/(sizeInPercent/100.0f)) orientation:UIImageOrientationUp];
     };
     
 }
