+++ src/Catty/SetXBrick.m
@@ -46,7 +46,7 @@
         NSDebug(@"Performing: %@", self.description);
         double xPosition = [self.xPosition interpretDoubleForSprite:self.object];
         
-        self.object.position = CGPointMake(xPosition, self.object.yPosition);
+        self.object.position = CGPointMake((CGFloat)xPosition, self.object.yPosition);
     };
 }
