+++ src/Catty/UIColor+CatrobatUIColorExtensions.m
@@ -85,7 +85,7 @@
 
 + (UIColor*)lightOrangeColor
 {
-    return [UIColor colorWithRed:232.0/255.0 green:146.0f/255.0f blue:6.0f/255.0f alpha:1.0f];
+    return [UIColor colorWithRed:(CGFloat)(232.0/255.0) green:146.0f/255.0f blue:6.0f/255.0f alpha:1.0f];
 }
 
 + (UIColor*)lightBlueColor;
